Working towards matching slowly but surely
