--- lib/al/resource/GameDataHolderWriter.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "GameDataHolder.h"
-
-class GameDataHolderWriter : public GameDataHolder {
-    // ima be real i dont think theres anything in this class
-};
