--- lib/al/resource/GameProgressData.h
@@ -1,8 +1,0 @@
-#pragma once
-
-class GameProgressData {
-public:
-    int getHomeLevel(void) const;
-    void upHomeLevel(void);
-    void talkCapNearHomeInWaterfall(void);
-};
