--- lib/game/Player.h
@@ -1,13 +1,0 @@
-#pragma once
-
-class IUseDimension;
-
-namespace al {
-class LiveActor;
-}
-
-namespace rs {
-
-bool isGuardNosePainCap(const al::LiveActor*);
-
-}  // namespace rs
