+++ lib/game/gameData/GameDataHolderBase.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace al {
+class ISceneObj {};
+};  // namespace al
+
+#include "al/hio/HioNode.h"
+#include "al/message/IUseMessageSystem.h"
+
+class GameDataHolderBase : public al::ISceneObj, public al::IUseMessageSystem, public al::HioNode {
+public:
+};
