+++ lib/game/gameData/GameDataHolderWriter.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "GameDataHolder.h"
+
+class GameDataHolderWriter : public GameDataHolder {
+    // ima be real i dont think theres anything in this class
+};
