+++ lib/game/gameData/progression/Achievement.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "game/gameData/GameDataHolderAccessor.h"
+
+class AchievementInfo;
+
+class Achievement {
+public:
+    Achievement(const AchievementInfo*);
+    bool isGet(GameDataHolderAccessor) const;
+
+private:
+    bool mGet = false;
+    const AchievementInfo* mInfo;
+};
