+++ lib/game/gameData/progression/GameProgressData.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class GameProgressData {
+public:
+    int getHomeLevel(void) const;
+    void upHomeLevel(void);
+    void talkCapNearHomeInWaterfall(void);
+};
