+++ lib/game/scenes/FirstSequenceScene/FirstSequenceScene.h
@@ -6,7 +6,7 @@ class FirstSequenceScene : public al::Scene {
     FirstSequenceScene(const char* name);
     ~FirstSequenceScene();
 
-    void init(const al::SceneInitInfo&);
+    void init(const al::SceneInitInfo&) override;
     void appear() override;
     void control() override;
     void drawMain() override;
