+++ lib/rs/Layout.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace al {
+class Scene;
+}
+namespace rs {
+void updateKitListLayoutOnlyLayoutScene(al::Scene* scene);
+}
