+++ src/al/resource/CMakeLists.txt
@@ -1,7 +1,3 @@
 target_sources(odyssey PRIVATE
-    Achievement.cpp
-    AchievementHolder.cpp
-    AchievementInfoReader.cpp
-    GameDataHolderAccessor.cpp
     Resource.cpp
 )
