+++ src/game/gameData/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_subdirectory(progression)
+
+target_sources(odyssey PRIVATE
+    GameDataHolderAccessor.cpp
+)
