+++ src/game/gameData/progression/Achievement.cpp
@@ -1,0 +1,10 @@
+#include "game/gameData/progression/Achievement.h"
+
+#include "game/gameData/progression/AchievementInfoReader.h"
+#include "al/util/ResourceUtil.h"
+
+Achievement::Achievement(const AchievementInfo* info) : mInfo(info) {}
+
+bool Achievement::isGet(GameDataHolderAccessor accessor) const {
+    return mGet || rs::checkGetAchievement(accessor, mInfo->mName);
+}
