+++ src/game/gameData/progression/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(odyssey PRIVATE
+    Achievement.cpp
+    AchievementHolder.cpp
+    AchievementInfoReader.cpp
+)
