+++ src/game/scenes/FirstSequenceScene/SceneStateSelectMode.cpp
@@ -1,7 +1,10 @@
 #include "game/scenes/FirstSequenceScene/SceneStateSelectMode.h"
 #include "al/input/KeyRepeatCtrl.h"
 #include "al/layout/actors/SimpleLayoutAppearWaitEnd.h"
+#include "al/rumble/WaveVibrationPlayer.h"
+#include "al/util/LayoutActorUtil.h"
 #include "rs/KidsMode.h"
+#include "rs/Layout.h"
 
 SceneStateSelectMode::SceneStateSelectMode(const char* name, al::Scene* scene,
                                            const al::LayoutInitInfo& initInfo,
@@ -14,20 +17,35 @@ SceneStateSelectMode::SceneStateSelectMode(const char* name, al::Scene* scene,
     mModeSelection->kill();
     mKeyRepeatCtrl = new al::KeyRepeatCtrl();
     mKeyRepeatCtrl->init(30, 5);
+    mNormalButton = new al::LayoutActor("モード選択[通常]");
+    al::initLayoutPartsActor(mNormalButton, mModeSelection, initInfo, "ParMode00", nullptr);
+    mKidsButton = new al::LayoutActor("モード選択[キッズ]");
+    al::initLayoutPartsActor(mKidsButton, mModeSelection, initInfo, "ParMode01", nullptr);
+    mSelectionCursor = new al::SimpleLayoutAppearWaitEnd(mModeSelection, "モード選択[カーソル]",
+                                                         "ParCursor", initInfo, nullptr);
+    al::killLayoutIfActive(mSelectionCursor);
 }
 
 void SceneStateSelectMode::appear() {
-    mIsDead = true;
+    mIsDead = false;
     al::setNerve(this, &nrvSceneStateSelectModeAppear);
-    rs::isKidsMode()
+    mIsNormal = !rs::isKidsMode(mKidsButton);
 }
 
 void SceneStateSelectMode::exeAppear() {
     if (al::isFirstStep(this)) {
         mModeSelection->appear();
+        al::hidePane(mModeSelection, "ParCursor");
+    }
 
+    rs::updateKitListLayoutOnlyLayoutScene(mHost);
+
+    if (!mModeSelection->isWait()) {
+        al::setNerve(this, &nrvSceneStateSelectModeWait);
     }
 }
+
+void SceneStateSelectMode::exeWait() {}
 
 namespace {
 NERVE_IMPL(SceneStateSelectMode, Appear)
@@ -36,4 +54,4 @@ NERVE_IMPL(SceneStateSelectMode, Decide)
 NERVE_IMPL(SceneStateSelectMode, DecideWait)
 NERVE_IMPL(SceneStateSelectMode, DecideConfirm)
 NERVE_IMPL(SceneStateSelectMode, End)
-}
+}  // namespace
